Create a view
Rails expects actions to have views associated with them to
display information.  It will first look for a template in
articles/new and if one is not found, it will then look in
application/new.  This is because the ArticlesController
inherits from ApplicationController. The two extensions for
the view indicate the format of the template and the handler
used to render the template.