Add links
The link_to method is one of Rails' built-in helpers that
creates a hyperlink based on the text to display and where
to go.  The welcome page specifies the controller to point
to because it is a different controller.  The new and show
pages don't need to specify the controller since they are
both using the same controller.