fix: Turbo issues
https://stackoverflow.com/questions/72100292/
rails-7-destroy-not-working-and-only-running-show

Found that Turbo scans the tabs in the
application for the turbo-method data
attribute, overriding the default GET
action.  Solution suggested fixed the
problem.

https://stackoverflow.com/questions/70606742/
rendering-a-view-after-form-submission-not-
working-correctly-in-rails-7

Found that Turbo expects a form submission
to be followed by a 303 code redirect. A
form submission would need to to rerender
the previous page again for unprocessable
articles.